requirements.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==5.3.1
certifi==2020.4.5.1
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.3
configparser==5.0.0
crayons==0.3.0
crcmod==1.7
cryptography==41.0.3
distlib==0.3.7
filelock==3.12.3
idna==2.9
jmespath==0.10.0
Markdown==3.4.4
markdown-it-py==3.0.0
mdurl==0.1.2
model-index==0.1.11
numpy==1.25.2
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.23
ordered-set==4.1.0
oss2==2.17.0
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.3.0
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.16.1
pyproject-api==1.6.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.23.0
rich==13.4.2
selenium==3.141.0
six==1.16.0
soupsieve==2.0.1
tabulate==0.9.0
tox==4.11.1
tqdm==4.65.2
tzdata==2023.3
urllib3==1.25.9
virtualenv==20.24.4
webdriver-manager==2.5.3
